meson: bump libqmi requirement to 1.37.95
Latest libqmi adds QMI_NAS_RADIO_INTERFACE_NO_CHANGE which will
cause failures when building with -Wall if NO_CHANGE isn't handled
in switch() statements. So we add that. But then of course you can't
build with older libqmi because it doesn't have NO_CHANGE.

So just bump the requirement.

Signed-off-by: Dan Williams <dan@ioncontrol.co>